Merge develop for 1.2.0 Update

+++ regolith/export.go
@@ -12,83 +12,33 @@ import (
 // resource pack based on exportTarget (a structure with data related to
 // export settings) and the name of the project.
 func GetExportPaths(
-	exportTarget ExportTarget, name string,
+	exportTarget ExportTarget, ctx RunContext,
 ) (bpPath string, rpPath string, err error) {
+	bpName, rpName, err := GetExportNames(exportTarget, ctx)
 	if exportTarget.Target == "development" {
 		comMojang, err := FindMojangDir()
 		if err != nil {
 			return "", "", burrito.WrapError(
 				err, "Failed to find \"com.mojang\" directory.")
 		}
-
-		// TODO - I don't like the _rp and _bp suffixes. Can we get rid of that?
-		// I for example always name my packs "0".
-		bpPath = comMojang + "/development_behavior_packs/" + name + "_bp"
-		rpPath = comMojang + "/development_resource_packs/" + name + "_rp"
+		return GetDevelopmentExportPaths(bpName, rpName, comMojang)
 	} else if exportTarget.Target == "preview" {
 		comMojang, err := FindPreviewDir()
 		if err != nil {
 			return "", "", burrito.WrapError(
 				err, "Failed to find preview \"com.mojang\" directory.")
 		}
-
-		// TODO - I don't like the _rp and _bp suffixes. Can we get rid of that?
-		// I for example always name my packs "0".
-		bpPath = comMojang + "/development_behavior_packs/" + name + "_bp"
-		rpPath = comMojang + "/development_resource_packs/" + name + "_rp"
+		return GetDevelopmentExportPaths(bpName, rpName, comMojang)
 	} else if exportTarget.Target == "exact" {
-		bpPath, err = ResolvePath(exportTarget.BpPath)
-		if err != nil {
-			return "", "", burrito.WrapError(
-				err, "Failed to resolve behavior pack path.")
-		}
-		rpPath, err = ResolvePath(exportTarget.RpPath)
-		if err != nil {
-			return "", "", burrito.WrapError(
-				err, "Failed to resolve resource pack path.")
-		}
+		return GetExactExportPaths(exportTarget)
 	} else if exportTarget.Target == "world" {
-		if exportTarget.WorldPath != "" {
-			if exportTarget.WorldName != "" {
-				return "", "", burrito.WrappedError(
-					"Using both \"worldName\" and \"worldPath\" is not" +
-						" allowed.")
-			}
-			wPath, err := ResolvePath(exportTarget.WorldPath)
-			if err != nil {
-				return "", "", burrito.WrapError(
-					err, "Failed to resolve world path.")
-			}
-			bpPath = filepath.Join(
-				wPath, "behavior_packs", name+"_bp")
-			rpPath = filepath.Join(
-				wPath, "resource_packs", name+"_rp")
-		} else if exportTarget.WorldName != "" {
-			dir, err := FindMojangDir()
-			if err != nil {
-				return "", "", burrito.WrapError(
-					err, "Failed to find \"com.mojang\" directory.")
-			}
-			worlds, err := ListWorlds(dir)
-			if err != nil {
-				return "", "", burrito.WrapError(err, "Failed to list worlds.")
-			}
-			for _, world := range worlds {
-				if world.Name == exportTarget.WorldName {
-					bpPath = filepath.Join(
-						world.Path, "behavior_packs", name+"_bp")
-					rpPath = filepath.Join(
-						world.Path, "resource_packs", name+"_rp")
-				}
-			}
-		} else {
-			err = burrito.WrappedError(
-				"The \"world\" export target requires either a " +
-					"\"worldName\" or \"worldPath\" property")
-		}
+		return GetWorldExportPaths(exportTarget, bpName, rpName)
 	} else if exportTarget.Target == "local" {
-		bpPath = "build/BP/"
-		rpPath = "build/RP/"
+		bpPath = "build/" + bpName + "/"
+		rpPath = "build/" + rpName + "/"
+	} else if exportTarget.Target == "none" {
+		bpPath = ""
+		rpPath = ""
 	} else {
 		err = burrito.WrappedErrorf(
 			"Export target %q is not valid", exportTarget.Target)
@@ -96,19 +46,121 @@ func GetExportPaths(
 	return
 }
 
+func GetDevelopmentExportPaths(bpName, rpName, comMojang string) (bpPath string, rpPath string, err error) {
+	if err != nil {
+		return "", "", burrito.WrapError(
+			err, "Failed to find \"com.mojang\" directory.")
+	}
+
+	bpPath = comMojang + "/development_behavior_packs/" + bpName
+	rpPath = comMojang + "/development_resource_packs/" + rpName
+	return
+}
+
+func GetExactExportPaths(exportTarget ExportTarget) (bpPath string, rpPath string, err error) {
+	bpPath, err = ResolvePath(exportTarget.BpPath)
+	if err != nil {
+		return "", "", burrito.WrapError(
+			err, "Failed to resolve behavior pack path.")
+	}
+	rpPath, err = ResolvePath(exportTarget.RpPath)
+	if err != nil {
+		return "", "", burrito.WrapError(
+			err, "Failed to resolve resource pack path.")
+	}
+	return
+}
+
+func GetWorldExportPaths(exportTarget ExportTarget, bpName, rpName string) (bpPath string, rpPath string, err error) {
+	if exportTarget.WorldPath != "" {
+		if exportTarget.WorldName != "" {
+			return "", "", burrito.WrappedError(
+				"Using both \"worldName\" and \"worldPath\" is not" +
+					" allowed.")
+		}
+		wPath, err := ResolvePath(exportTarget.WorldPath)
+		if err != nil {
+			return "", "", burrito.WrapError(
+				err, "Failed to resolve world path.")
+		}
+		bpPath = filepath.Join(
+			wPath, "behavior_packs", bpName)
+		rpPath = filepath.Join(
+			wPath, "resource_packs", rpName)
+	} else if exportTarget.WorldName != "" {
+		dir, err := FindMojangDir()
+		if err != nil {
+			return "", "", burrito.WrapError(
+				err, "Failed to find \"com.mojang\" directory.")
+		}
+		worlds, err := ListWorlds(dir)
+		if err != nil {
+			return "", "", burrito.WrapError(err, "Failed to list worlds.")
+		}
+		for _, world := range worlds {
+			if world.Name == exportTarget.WorldName {
+				bpPath = filepath.Join(
+					world.Path, "behavior_packs", bpName)
+				rpPath = filepath.Join(
+					world.Path, "resource_packs", rpName)
+			}
+		}
+	} else {
+		err = burrito.WrappedError(
+			"The \"world\" export target requires either a " +
+				"\"worldName\" or \"worldPath\" property")
+	}
+	return
+}
+
+// GetExportNames returns the names for the behavior pack and resource pack
+// based on the evaluated values of the "bpName" and "rpName" from the
+// exportTarget object.
+func GetExportNames(exportTarget ExportTarget, ctx RunContext) (bpName string, rpName string, err error) {
+	if exportTarget.BpName != "" {
+		bpName, err = EvalString(exportTarget.BpName, ctx)
+		if err != nil {
+			return "", "", burrito.WrapError(
+				err, "Failed to evaluate behavior pack name.")
+		}
+	} else {
+		bpName = ctx.Config.Name + "_bp"
+	}
+	if exportTarget.RpName != "" {
+		rpName, err = EvalString(exportTarget.RpName, ctx)
+		if err != nil {
+			return "", "", burrito.WrapError(
+				err, "Failed to evaluate resource pack name.")
+		}
+	} else {
+		rpName = ctx.Config.Name + "_rp"
+	}
+	return
+}
+
 // ExportProject copies files from the tmp paths (tmp/BP and tmp/RP) into
 // the project's export target. The paths are generated with GetExportPaths.
-func ExportProject(
-	profile Profile, name, dataPath, dotRegolithPath string,
-) error {
+func ExportProject(ctx RunContext) error {
+	MeasureStart("Export - GetExportPaths")
+	profile, err := ctx.GetProfile()
+	if err != nil {
+		return burrito.WrapError(err, runContextGetProfileError)
+	}
+	if profile.ExportTarget.Target == "none" {
+		Logger.Debugf("Export target is set to \"none\". Skipping export.")
+		return nil
+	}
+	dataPath := ctx.Config.DataPath
+	dotRegolithPath := ctx.DotRegolithPath
 	// Get the export target paths
 	exportTarget := profile.ExportTarget
-	bpPath, rpPath, err := GetExportPaths(exportTarget, name)
+	bpPath, rpPath, err := GetExportPaths(exportTarget, ctx)
 	if err != nil {
 		return burrito.WrapError(
 			err, "Failed to get generate export paths.")
 	}
 
+	MeasureStart("Export - LoadEditedFiles")
 	// Loading edited_files.json or creating empty object
 	editedFiles := LoadEditedFiles(dotRegolithPath)
 	err = editedFiles.CheckDeletionSafety(rpPath, bpPath)
@@ -124,22 +176,28 @@ func ExportProject(
 			rpPath, bpPath)
 	}
 
-	// Clearing output locations
-	// Spooky, I hope file protection works, and it won't do any damage
-	err = os.RemoveAll(bpPath)
-	if err != nil {
-		return burrito.WrapErrorf(
-			err, "Failed to clear behavior pack from build path %q.\n"+
-				"Are user permissions correct?", bpPath)
+	MeasureStart("Export - Clean")
+	// When comparing the size and modification time of the files, we need to
+	// keep the files in target paths.
+	if !IsExperimentEnabled(SizeTimeCheck) {
+		// Clearing output locations
+		// Spooky, I hope file protection works, and it won't do any damage
+		err = os.RemoveAll(bpPath)
+		if err != nil {
+			return burrito.WrapErrorf(
+				err, "Failed to clear behavior pack from build path %q.\n"+
+					"Are user permissions correct?", bpPath)
+		}
+		err = os.RemoveAll(rpPath)
+		if err != nil {
+			return burrito.WrapErrorf(
+				err, "Failed to clear resource pack from build path %q.\n"+
+					"Are user permissions correct?", rpPath)
+		}
 	}
-	err = os.RemoveAll(rpPath)
-	if err != nil {
-		return burrito.WrapErrorf(
-			err, "Failed to clear resource pack from build path %q.\n"+
-				"Are user permissions correct?", rpPath)
-	}
+	MeasureEnd()
 	// List the names of the filters that opt-in to the data export process
-	exportedFilterNames := []string{}
+	var exportedFilterNames []string
 	for filter := range profile.Filters {
 		filter := profile.Filters[filter]
 		usingDataPath, err := filter.IsUsingDataExport(dotRegolithPath)
@@ -175,11 +233,10 @@ func ExportProject(
 			err1 = os.MkdirAll(dataPath, 0755)
 		}
 		if err1 != nil {
-			return burrito.WrapErrorf(
-				err, "Failed to read the files from the data path %q",
-				dataPath)
+			return burrito.WrapErrorf(err, osReadDirError, dataPath)
 		}
 	}
+	MeasureStart("Export - RevertibleOps")
 	// Create revertible operations object
 	backupPath := filepath.Join(dotRegolithPath, ".dataBackup")
 	revertibleOps, err := NewRevertibleFsOperations(backupPath)
@@ -187,6 +244,7 @@ func ExportProject(
 		return burrito.WrapErrorf(err, newRevertibleFsOperationsError, backupPath)
 	}
 	// Export data
+	MeasureStart("Export - ExportData")
 	for _, exportedFilterName := range exportedFilterNames {
 		// Clear export target
 		targetPath := filepath.Join(dataPath, exportedFilterName)
@@ -230,18 +288,35 @@ func ExportProject(
 			return mainError
 		}
 	}
-	// Export BP
-	Logger.Infof("Exporting behavior pack to \"%s\".", bpPath)
-	err = MoveOrCopy(filepath.Join(dotRegolithPath, "tmp/BP"), bpPath, exportTarget.ReadOnly, true)
-	if err != nil {
-		return burrito.WrapError(err, "Failed to export behavior pack.")
+	MeasureStart("Export - MoveOrCopy")
+	if IsExperimentEnabled(SizeTimeCheck) {
+		// Export BP
+		Logger.Infof("Exporting behavior pack to \"%s\".", bpPath)
+		err = SyncDirectories(filepath.Join(dotRegolithPath, "tmp/BP"), bpPath, exportTarget.ReadOnly)
+		if err != nil {
+			return burrito.WrapError(err, "Failed to export behavior pack.")
+		}
+		// Export RP
+		Logger.Infof("Exporting project to \"%s\".", filepath.Clean(rpPath))
+		err = SyncDirectories(filepath.Join(dotRegolithPath, "tmp/RP"), rpPath, exportTarget.ReadOnly)
+		if err != nil {
+			return burrito.WrapError(err, "Failed to export resource pack.")
+		}
+	} else {
+		// Export BP
+		Logger.Infof("Exporting behavior pack to \"%s\".", bpPath)
+		err = MoveOrCopy(filepath.Join(dotRegolithPath, "tmp/BP"), bpPath, exportTarget.ReadOnly, true)
+		if err != nil {
+			return burrito.WrapError(err, "Failed to export behavior pack.")
+		}
+		// Export RP
+		Logger.Infof("Exporting project to \"%s\".", filepath.Clean(rpPath))
+		err = MoveOrCopy(filepath.Join(dotRegolithPath, "tmp/RP"), rpPath, exportTarget.ReadOnly, true)
+		if err != nil {
+			return burrito.WrapError(err, "Failed to export resource pack.")
+		}
 	}
-	// Export RP
-	Logger.Infof("Exporting project to \"%s\".", filepath.Clean(rpPath))
-	err = MoveOrCopy(filepath.Join(dotRegolithPath, "tmp/RP"), rpPath, exportTarget.ReadOnly, true)
-	if err != nil {
-		return burrito.WrapError(err, "Failed to export resource pack.")
-	}
+	MeasureStart("Export - UpdateFromPaths")
 	// Update or create edited_files.json
 	err = editedFiles.UpdateFromPaths(rpPath, bpPath)
 	if err != nil {
@@ -258,6 +333,7 @@ func ExportProject(
 	if err := revertibleOps.Close(); err != nil {
 		return burrito.PassError(err)
 	}
+	MeasureEnd()
 	return nil
 }
